+++ packages/cloud_firestore/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.12.5
+
+* Makes `startAtDocument`, `startAfterDocument`, `endAtDocument` and `endBeforeDocument` work
+  with `Query.collectionGroup` queries.
+* Fixes `startAtDocument`, `startAfterDocument`, `endAtDocument` and `endBeforeDocument` to
+  also work with a descending order as the last explicit sort order.
+* Fixed an integration test by increasing the value of `cacheSizeBytes` to a valid value.
+
 ## 0.12.4
 
 * Added support for `Query.collectionGroup`.